requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
numpy>=1.20.0
pandas>=1.3.0
openpyxl>=3.0.9
PyJWT>=2.1.0
geojson>=2.5.0
python-dotenv>=0.19.0
pycryptodome>=3.23.0